+++ CHANGELOG.md
@@ -11,6 +11,10 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- Make `ScaleJitter` in LTDETR step-aware. Now you can stop the augmentation by adding a
+  `step_stop` args like the following
+  `transform_args={"scale_jitter": {"step_stop": 10000}}`
+
 ### Deprecated
 
 ### Removed